+++ src/mha/returns/annually.py
@@ -31,7 +31,7 @@ def annual_time_weighted_returns(
 
     return float(np.dot(weights, r))
 
-def find_annaul_estimations(symbol: str, decay_parameter: float | None = None, lookback: float | None = None) -> np.ndarray:
+def find_annaul_estimations(symbol: str, decay_parameter: float | None = None, lookback: int | None = None) -> np.ndarray:
     #  Step 1-- Data Ingestion
 
     #    Data Based on horizon is loaded and cleaned
